feat(admin): 관리자 정보 변경 비지니스 로직 구현(#21)
/api/v1/admins/{adminId} PATCH 경로의 api를 구현했습니다.

+++ src/main/java/com/yeo_li/yeol_post/category/CategoryRepository.java
@@ -8,5 +8,5 @@ public interface CategoryRepository extends JpaRepository<Category, Long> {
 
   Category findCategoryByCategoryName(String categoryName);
 
-  Category findCategoryById(int categoryId);
+  Category findCategoryById(Long categoryId);
 }
